requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.2
numpy==1.16.1
opencv-python==3.1.0.4
psycopg2-binary==2.8.3
PyJWT==1.5.3
pymongo==3.6.1
pyparsing==2.3.1
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2017.3
requests==2.18.4
six==1.11.0
soupsieve==1.9.1
twilio==6.10.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.2.1
